LayoutTests/fast/css-grid-layout/named-grid-line-get-set-expected.txt

Test that setting and getting grid-definition-columns and grid-definition-rows works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting -webkit-grid-definition-columns and -webkit-grid-definition-rows set through CSS
PASS window.getComputedStyle(gridWithFixedElement, '').getPropertyValue('-webkit-grid-definition-columns') is "first 10px"
PASS window.getComputedStyle(gridWithFixedElement, '').getPropertyValue('-webkit-grid-definition-rows') is "first 15px"
PASS window.getComputedStyle(gridWithPercentElement, '').getPropertyValue('-webkit-grid-definition-columns') is "424px last"
PASS window.getComputedStyle(gridWithPercentElement, '').getPropertyValue('-webkit-grid-definition-rows') is "162px last"
PASS window.getComputedStyle(gridWithAutoElement, '').getPropertyValue('-webkit-grid-definition-columns') is "first 0px"
PASS window.getComputedStyle(gridWithAutoElement, '').getPropertyValue('-webkit-grid-definition-rows') is "0px last"
PASS window.getComputedStyle(gridWithAutoWithChildrenElement, '').getPropertyValue('-webkit-grid-definition-columns') is "first 77px"
PASS window.getComputedStyle(gridWithAutoWithChildrenElement, '').getPropertyValue('-webkit-grid-definition-rows') is "22px last"
PASS window.getComputedStyle(gridWithMinMax, '').getPropertyValue('-webkit-grid-definition-columns') is "first 80px"
PASS window.getComputedStyle(gridWithMinMax, '').getPropertyValue('-webkit-grid-definition-rows') is "300px last"
PASS window.getComputedStyle(gridWithFixedMultiple, '').getPropertyValue('-webkit-grid-definition-columns') is "nav first 10px last"
PASS window.getComputedStyle(gridWithFixedMultiple, '').getPropertyValue('-webkit-grid-definition-rows') is "nav first 15px last"
PASS window.getComputedStyle(gridWithPercentageSameStringMultipleTimes, '').getPropertyValue('-webkit-grid-definition-columns') is "nav first 80px nav 120px last"
PASS window.getComputedStyle(gridWithPercentageSameStringMultipleTimes, '').getPropertyValue('-webkit-grid-definition-rows') is "first nav2 150px nav2 450px last"

Test getting and setting -webkit-grid-definition-columns and -webkit-grid-definition-rows through JS
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "first 18px"
PASS element.style.webkitGridDefinitionColumns is "first 18px"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "66px last"
PASS element.style.webkitGridDefinitionRows is "66px last"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "first 440px"
PASS element.style.webkitGridDefinitionColumns is "first 55%"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "240px last"
PASS element.style.webkitGridDefinitionRows is "40% last"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "first 0px"
PASS element.style.webkitGridDefinitionColumns is "first auto"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "0px last"
PASS element.style.webkitGridDefinitionRows is "auto last"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "first 0px"
PASS element.style.webkitGridDefinitionColumns is "first -webkit-min-content"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "0px last"
PASS element.style.webkitGridDefinitionRows is "-webkit-min-content last"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "first 0px"
PASS element.style.webkitGridDefinitionColumns is "first -webkit-max-content"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "0px last"
PASS element.style.webkitGridDefinitionRows is "-webkit-max-content last"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "first 440px"
PASS element.style.webkitGridDefinitionColumns is "first minmax(55%, 45px)"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "240px last"
PASS element.style.webkitGridDefinitionRows is "minmax(30px, 40%) last"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "first 220px"
PASS element.style.webkitGridDefinitionColumns is "first minmax(22em, -webkit-max-content)"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "50px last"
PASS element.style.webkitGridDefinitionRows is "minmax(-webkit-max-content, 5em) last"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "first 220px"
PASS element.style.webkitGridDefinitionColumns is "first minmax(22em, -webkit-min-content)"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "50px last"
PASS element.style.webkitGridDefinitionRows is "minmax(-webkit-min-content, 5em) last"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "first 0px"
PASS element.style.webkitGridDefinitionColumns is "first minmax(-webkit-min-content, -webkit-max-content)"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "0px last"
PASS element.style.webkitGridDefinitionRows is "minmax(-webkit-max-content, -webkit-min-content) last"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "nav first 0px last"
PASS element.style.webkitGridDefinitionColumns is "first nav minmax(-webkit-min-content, -webkit-max-content) last"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "nav first 0px last"
PASS element.style.webkitGridDefinitionRows is "first nav minmax(-webkit-max-content, -webkit-min-content) last"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "nav first 0px nav 0px last"
PASS element.style.webkitGridDefinitionColumns is "first nav minmax(-webkit-min-content, -webkit-max-content) nav auto last"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "first nav2 0px nav2 15px last"
PASS element.style.webkitGridDefinitionRows is "first nav2 minmax(-webkit-max-content, -webkit-min-content) nav2 minmax(10px, 15px) last"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "foo bar 0px foo 0px bar"
PASS element.style.webkitGridDefinitionColumns is "foo bar auto foo auto bar"
PASS getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "foo bar 0px foo 0px bar"
PASS element.style.webkitGridDefinitionRows is "foo bar auto foo auto bar"

Test getting and setting invalid -webkit-grid-definition-columns and -webkit-grid-definition-rows through JS
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-columns') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('-webkit-grid-definition-rows') is "none"
PASS successfullyParsed is true

TEST COMPLETE
